# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Proyecto Gimnasio - HU09 y HU10",
    "version": "1.0.0",
    "description": "Contrato backend de recuperaci?n de credenciales, objetivos y operaciones existentes afectadas."
  },
  "components": {
    "securitySchemes": {
      "session": {
        "type": "apiKey",
        "in": "cookie",
        "name": "gym_session"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "error": {
            "type": "string"
          }
        },
        "required": ["error"]
      },
...
    }
  },
  "paths": {
...
    "/auth/password-reset": {
      "post": {
        "operationId": "resetPassword",
        "summary": "Restablecer contrase?a con token",
        "description": "Consume el token una sola vez, verifica vencimiento en el servidor y revoca todas las sesiones y enlaces previos dentro de la transacci?n. No cambia roles ni gimnasio.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": false,
                "properties": {
                  "token": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 256,
                    "writeOnly": true
                  },
                  "password": {
                    "type": "string",
                    "minLength": 8,
                    "maxLength": 72,
                    "writeOnly": true,
                    "description": "Misma pol?tica HU06: may?scula, min?scula y n?mero. M?ximo 72 bytes UTF-8 por bcrypt."
                  }
                },
                "required": ["token", "password"]
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Contrase?a actualizada; todas las sesiones previas revocadas."
          },
          "400": {
            "description": "weak_password o invalid_recovery_token gen?rico para token vencido, usado, reemplazado o inexistente.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  }
}